Avoid conflict with ntp in Debian Bullseye
build-iso for Debian 11 results in unresolved build-dependencies:

conflict for providers of systemd-timesyncd needed by live-build-desc-bullseye, (provider systemd-timesyncd is in conflict with ntp)

The bullseye.deps in live-build-desc is based on the debootstrap essential packages
for Debian Bullseye. systemd-timesyncd is preferred over ntp in Debian 11.

To keep our NTP client glue-code working, we need to adjust the essential package
lists in the live-build-desc.

VRVDR-56508